requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
pillow
sqlparse
tzdata
djangorestframework-simplejwt